Add explicit configuration for php-fpm
If the current `php_enable_php_fpm` variable is set to true, php-fpm
will be both configured and started. This may not be the desired outcome
for those who want to configure php-fpm without starting the process.

Add a new variable, `php_start_php_fpm`, that can be configured to
handle the process behavior. This will set to the value of
`php_enable_php_fpm` by default to maintain backwards compatibility,
but can be modified as desired.