Adding support for Ondrej's PHP PPA for Ubuntu OS to this playbook. It is possible to install the ppa using the variable
	php_ppa_use: true

It is also now possible to change the PHP version to install under Ubuntu by specifying the variable

	php_version: "7.1"

As left default version to the Debian version which is at the moment 7.0.
I wrote the change in this playbook using the same strategy that has been used in the role `ansible-role-nginx`.