Add php_fpm_clear_env parameter
Using the clear_env you can control if a FPM process will receive the
default environment variables[^1] such as the Path. For example nextcloud
requires / recommends the path variable for some operations[^2], therefore I
would assume this to be widely used enough to put it directly in the
core of the role.

[^1]: <https://www.php.net/manual/en/install.fpm.configuration.php#clear-env>
[^2]: <https://docs.nextcloud.com/server/19/admin_manual/installation/source_installation.html#php-fpm-configuration-notes>